pyproject.toml
[tool.poetry]
name = "algot_market_data"
version = "0.1.0"
description = ""
authors = ["Łukasz Sawaniewski <l.sawaniewski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.18.0"
websocket-client = "^1.1.0"
requests = "^2.26.0"
pandas = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
